refactor(operation_funcs.py -> operations.py)
