+++ automata/operations.py
