+++ tests/unit/regular_operations/test_conversion.py
@@ -1,6 +1,6 @@
 from automata.automaton import Epsilon
 from automata.dfa import DFA
-from automata.operation_funcs import convert_nfa_to_dfa
+from automata.operations import convert_nfa_to_dfa
 from tests.conftest import make_nfa
