+++ tests/unit/regular_operations/test_union.py
@@ -1,5 +1,5 @@
 from automata.automaton import Epsilon
-from automata.operation_funcs import union
+from automata.operations import union
 from tests.conftest import make_nfa
 
 # ───────────────────────────────
